requirements.txt
opencv-python==4.8.0.74
numpy==1.23.5
python-dotenv==1.0.0
scikit-learn==1.2.2
pdf2image==1.16.3
google-generativeai==0.4.1
Pillow==10.0.1
